+++ CHANGELOG.rst
@@ -2,6 +2,7 @@
 ---------
 
 * Adds full categorization of Emojis
+* Adds update link to forum template generator
 
 2020-11-1
 ---------